# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance Management API",
        "description": "API permettant de gérer les transactions financières des utilisateurs",
        "contact": {
            "email": "support@financeapi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer toutes les transactions de l'utilisateur connecté",
                "operationId": "getTransactions",
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une nouvelle transaction",
                "operationId": "createTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "amount",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salaire"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "earning"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1000.50"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-04"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer une transaction spécifique",
                "operationId": "getTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Mettre à jour une transaction existante",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "amount",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salaire"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "earning"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1000.50"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-04"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Supprimer une transaction",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction supprimée avec succès"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/earnings": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer les revenus de l'utilisateur connecté",
                "operationId": "getEarnings",
                "responses": {
                    "200": {
                        "description": "Liste des revenus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/earnings-sum": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer la somme des revenus de l'utilisateur connecté",
                "operationId": "getEarningsSum",
                "responses": {
                    "200": {
                        "description": "Somme des revenus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_earnings": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "1500.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/expenses": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer les dépenses de l'utilisateur connecté",
                "operationId": "getExpenses",
                "responses": {
                    "200": {
                        "description": "Liste des dépenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/expenses-sum": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer la somme des dépenses de l'utilisateur connecté",
                "operationId": "getExpensesSum",
                "responses": {
                    "200": {
                        "description": "Somme des dépenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_expenses": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "500.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/total-sum": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer la somme totale des transactions de l'utilisateur connecté",
                "operationId": "getTotalSum",
                "responses": {
                    "200": {
                        "description": "Somme totale",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sum": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "1000.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Transaction": {
                "title": "Transaction",
                "description": "Details of a financial transaction",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the transaction",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name or description of the transaction",
                        "type": "string",
                        "example": "Transaction 1"
                    },
                    "user_id": {
                        "description": "The ID of the user associated with the transaction",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "The type of the transaction (e.g., earning, expense)",
                        "type": "string",
                        "example": "earning"
                    },
                    "amount": {
                        "description": "The amount of the transaction",
                        "type": "number",
                        "format": "float",
                        "example": 12
                    },
                    "date": {
                        "description": "The date of the transaction",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-08"
                    },
                    "created_at": {
                        "description": "The date and time when the transaction was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-04T13:19:47Z"
                    },
                    "updated_at": {
                        "description": "The date and time when the transaction was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-04T13:19:47Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}